Move low-level soundfont loading functions to separate files
